[UI] [Backend] potentially fix camera settings, various logging tweaks

+++ photon-server/src/main/java/org/photonvision/vision/pipeline/CVPipelineSettings.java
@@ -39,9 +39,9 @@ public class CVPipelineSettings {
     public ImageFlipMode inputImageFlipMode = ImageFlipMode.NONE;
     public ImageRotationMode inputImageRotationMode = ImageRotationMode.DEG_0;
     public String pipelineNickname = "New Pipeline";
-    public double cameraExposure = 50.0;
-    public double cameraBrightness = 50.0;
-    public double cameraGain = 50.0;
+    public int cameraExposure = 50;
+    public int cameraBrightness = 50;
+    public int cameraGain = 50;
     public int cameraVideoModeIndex = 0;
     public FrameDivisor inputFrameDivisor = FrameDivisor.NONE;
     public FrameDivisor outputFrameDivisor = FrameDivisor.NONE;
